Rename Function1 to AzureFunction
Removed the file Function1.cs and added a new file AzureFunction.cs. This change was made to better reflect the nature of the function, making the name more clear and aligned with its purpose in the context of Azure Functions. The content of the file remains the same with the only change being the name of the class, function and the logger instance.