Updated Azure Function to check prime numbers
The AzureFunction now checks whether a number is prime or not. The operation function name has been updated from "AzureFunction" to "CheckIfPrimeNumberOrNot". Added response section to return whether the supplied number is prime. Added NumberData class which acts as a model for the number to be checked. Added PrimeNumber class with 'IsPrime' static method that checks whether a number is prime. The trigger now accepts post requests and can be accessed by anonymous users.